+++ src/main/java/org/ow2/authzforce/core/pdp/api/AttributeProvider.java
@@ -19,7 +19,7 @@ package org.ow2.authzforce.core.pdp.api;
 
 import org.ow2.authzforce.core.pdp.api.value.AttributeBag;
 import org.ow2.authzforce.core.pdp.api.value.AttributeValue;
-import org.ow2.authzforce.core.pdp.api.value.BagDatatype;
+import org.ow2.authzforce.core.pdp.api.value.Datatype;
 
 /**
  * Attribute provider used to resolve {@link oasis.names.tc.xacml._3_0.core.schema.wd_17.AttributeDesignatorType}s in a specific way (e.g. from a specific attribute source)
@@ -34,13 +34,15 @@ public interface AttributeProvider
 	 * @param attributeFQN
 	 *            the global identifier (Category,Issuer,AttributeId) of the attribute to find
 	 * @param context
-	 *            the representation of the request data
-	 * @param returnDatatype
-	 *            attribute value bag datatype
-	 * @return the result of retrieving the attribute, which will be a bag of values of type defined by {@code attributeDatatype}; empty bag iff no value found and no error occurred.
+	 *            the request context
+	 * @param datatype
+	 *            attribute datatype
+	 * @return the result of retrieving the attribute, which will be a bag of values of type defined by {@code returnDatatype}; empty bag iff no value found and no error occurred.
+	 * @throws UnsupportedOperationException
+	 *             {@code attributeFQN} or {@code returnDatatype} are not supported (the PDP engine should try another attribute provider if any)
 	 * @throws IndeterminateEvaluationException
-	 *             if any error finding the attribute value(s)
+	 *             {@code attributeFQN} or {@code returnDatatype} are supported but some error occurred while trying to resolve the attribute value(s)
 	 */
-	<AV extends AttributeValue> AttributeBag<AV> get(AttributeFqn attributeFQN, BagDatatype<AV> returnDatatype, EvaluationContext context) throws IndeterminateEvaluationException;
+	<AV extends AttributeValue> AttributeBag<AV> get(AttributeFqn attributeFQN, Datatype<AV> datatype, EvaluationContext context) throws IndeterminateEvaluationException;
 
 }